requirements.txt
anyio==3.6.1
async-timeout==4.0.2
autopep8==1.6.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
click==8.1.3
Deprecated==1.2.13
facenet-pytorch==2.5.2
fastapi==0.79.0
google-api-core==2.2.2
google-auth==2.3.3
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
h11==0.13.0
idna==2.8
isodate==0.6.1
Jinja2==3.0.3
lxml==4.6.5
MarkupSafe==2.0.0
numpy==1.23.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
pendulum==2.1.2
Pillow==9.0.1
playsound==1.3.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
pyparsing==3.0.7
PyPDF2==1.26.0
python-dateutil==2.7.3
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2019.3
pytzdata==2020.1
redis==4.3.4
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2
starlette==0.19.1
toml==0.10.2
torch==1.12.0
torchvision==0.13.0
typing_extensions==4.3.0
urllib3==1.26.5
uvicorn==0.18.2
wrapt==1.14.1
zope.event==4.5.0
zope.interface==5.4.0
